+++ lib/feature/auth/presentation/page/reset_password_page.dart
@@ -2,32 +2,23 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:share_kakeibo/feature/login/data/login_repository_impl.dart';
-import 'package:share_kakeibo/feature/login/presentation/state/auth_state.dart';
+import 'package:share_kakeibo/feature/auth/application/auth_service.dart';
 import 'package:share_kakeibo/importer.dart';
 
-class RegisterPage extends HookConsumerWidget {
-  const RegisterPage({Key? key}) : super(key: key);
+class ResetPasswordPage extends HookConsumerWidget {
+  const ResetPasswordPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final loginNotifier = ref.watch(loginRepositoryProvider);
-    final scaffoldMessenger = ref.watch(scaffoldKeyProvider).currentState!;
+    // authのサービス
+    final authService = ref.watch(authServiceProvider);
 
-    // ユーザー名
-    final userName = useState('');
-    final userNameController = useState(TextEditingController());
+    // snackbarの設定
+    final scaffoldMessenger = ref.watch(scaffoldKeyProvider).currentState!;
 
     // メールアドレス
     final email = useState('');
     final emailController = useState(TextEditingController());
-
-    // パスワード
-    final password = useState('');
-    final passwordController = useState(TextEditingController());
-
-    // パスワードの非表示
-    final _isObscure = useState(true);
 
     return Scaffold(
       body: SafeArea(
@@ -38,39 +29,27 @@ class RegisterPage extends HookConsumerWidget {
               child: Column(
                 children: [
                   const Logo(),
-                  const SizedBox(height: 32),
-                  LoginTextField(
-                    labelText: 'ユーザー名',
-                    controller: userNameController.value,
-                    textChange: (text) => userName.value = text,
-                  ),
                   const SizedBox(height: 16),
+                  Text(
+                    'パスワード再設定メールを送信',
+                    style: context.bodyLargeBold,
+                  ),
+                  const SizedBox(height: 32),
                   LoginTextField(
                     labelText: 'メールアドレス',
                     controller: emailController.value,
                     textChange: (text) => email.value = text,
                   ),
                   const SizedBox(height: 16),
-                  LoginTextField(
-                    labelText: 'パスワード（6桁以上）',
-                    controller: passwordController.value,
-                    textChange: (text) => password.value = text,
-                    isObscure: _isObscure.value,
-                    isObscureChange: () => _isObscure.value = !_isObscure.value,
-                  ),
-                  const SizedBox(height: 16),
                   CustomElevatedButton(
-                    text: '新規登録',
+                    text: '送信',
                     onTaped: () async {
                       showProgressDialog(context);
                       try {
-                        // ユーザー名、メールアドレス、パスワードのバリデーション
-                        final validMessage = Validator.validateUserName(
-                                value: userName.value) ??
-                            Validator.validateEmail(value: email.value) ??
-                            Validator.validatePassword(value: password.value);
+                        // メールアドレスのバリデーション
+                        final validMessage =
+                            Validator.validateEmail(value: email.value);
                         if (validMessage != null) {
-                          Navigator.of(context).pop();
                           final snackbar = CustomSnackBar(
                             context,
                             msg: validMessage,
@@ -80,24 +59,22 @@ class RegisterPage extends HookConsumerWidget {
                           return;
                         }
 
-                        await loginNotifier.register(
-                          uidProvider: ref.read(uidProvider.notifier).state,
-                          userName: userName.value,
-                          email: email.value,
-                          password: password.value,
-                        );
+                        // メール送信処理
+                        await authService.resetPassword(email: email.value);
+
+                        // メール送信処理後、ログイン画面へ戻る
                         Navigator.popUntil(
                             context, (Route<dynamic> route) => route.isFirst);
                         final snackbar = CustomSnackBar(
                           context,
-                          msg: 'アカウントの新規登録が完了しました',
+                          msg: 'パスワード再設定メールを送信しました',
                         );
                         scaffoldMessenger.showSnackBar(snackbar);
                       } on FirebaseAuthException catch (e) {
                         Navigator.of(context).pop();
                         final snackbar = CustomSnackBar(
                           context,
-                          msg: ref.watch(authNotifierProvider).getErrorMessage(e),
+                          msg: authService.getErrorMessage(e),
                           color: Colors.red,
                         );
                         scaffoldMessenger.showSnackBar(snackbar);
@@ -111,7 +88,7 @@ class RegisterPage extends HookConsumerWidget {
                   const Divider(),
                   TextButton(
                     child: Text(
-                      '既にアカウントをお持ちの方はこちら',
+                      'ログイン画面へ戻る',
                       style: context.bodyMediumGrey,
                     ),
                     onPressed: () => Navigator.of(context).pop(),
