+++ lib/feature/auth/presentation/widget/logo.dart
