+++ lib/feature/room/presentation/page/room_info_page.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:share_kakeibo/common_widget/dialog/input_text_dialog.dart';
-import 'package:share_kakeibo/feature/login/data/login_repository_impl.dart';
+import 'package:share_kakeibo/feature/auth/application/auth_service.dart';
+import 'package:share_kakeibo/feature/auth/data/auth_repository_impl.dart';
 import 'package:share_kakeibo/importer.dart';
 
 class RoomInfoPage extends HookConsumerWidget {
@@ -61,7 +62,7 @@ class RoomInfoPage extends HookConsumerWidget {
                   return;
                 }
 
-                await ref.read(loginRepositoryProvider).reSingIn(
+                await ref.read(authServiceProvider).reSingIn(
                       email: userData.email,
                       password: password,
                     );
