+++ lib/feature/user/presentation/page/email_page.dart
@@ -3,7 +3,8 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:share_kakeibo/common_widget/custom_text_field.dart';
 import 'package:share_kakeibo/common_widget/dialog/input_text_dialog.dart';
-import 'package:share_kakeibo/feature/login/data/login_repository_impl.dart';
+import 'package:share_kakeibo/feature/auth/application/auth_service.dart';
+import 'package:share_kakeibo/feature/auth/data/auth_repository_impl.dart';
 import 'package:share_kakeibo/importer.dart';
 
 class EmailPage extends HookConsumerWidget {
@@ -86,7 +87,7 @@ class EmailPage extends HookConsumerWidget {
                         }
 
                         // ユーザー情報変更のために再ログイン
-                        await ref.read(loginRepositoryProvider).reSingIn(
+                        await ref.read(authServiceProvider).reSingIn(
                               email: userData!.email,
                               password: password,
                             );
