+++ lib/feature/user/presentation/page/password_page.dart
@@ -3,7 +3,8 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:share_kakeibo/common_widget/custom_text_field.dart';
 import 'package:share_kakeibo/common_widget/dialog/input_text_dialog.dart';
-import 'package:share_kakeibo/feature/login/data/login_repository_impl.dart';
+import 'package:share_kakeibo/feature/auth/application/auth_service.dart';
+import 'package:share_kakeibo/feature/auth/data/auth_repository_impl.dart';
 import 'package:share_kakeibo/feature/user/data/user_repository.impl.dart';
 import 'package:share_kakeibo/importer.dart';
 
@@ -78,7 +79,7 @@ class PasswordPage extends HookConsumerWidget {
                           return;
                         }
 
-                        await ref.read(loginRepositoryProvider).reSingIn(
+                        await ref.read(authServiceProvider).reSingIn(
                               email: userData!.email,
                               password: reSignInPwd,
                             );
