+++ lib/importer.dart
@@ -1,9 +1,9 @@
 export 'app.dart';
 export 'common_widget/custom_pie_chart.dart';
-export 'feature/login/presentation/widget/logo.dart';
+export 'feature/auth/presentation/widget/logo.dart';
 export 'feature/calendar/presentation/widget/calendar_event_list.dart';
 export 'feature/event/presentation/widget/upsert_event_item.dart';
-export 'feature/login/presentation/widget/login_text_field.dart';
+export 'feature/auth/presentation/widget/login_text_field.dart';
 export 'feature/calendar/presentation/widget/custom_calendar.dart';
 export 'feature/root_page.dart';
 export 'common_widget/custom_drawer.dart';
@@ -45,9 +45,9 @@ export 'util/logger.dart';
 export 'feature/calendar/presentation/page/calendar_page.dart';
 export 'feature/chart/presentation/page/chart_page.dart';
 export 'feature/home/presentation/page/home_page.dart';
-export 'feature/login/presentation/page/login_page.dart';
-export 'feature/login/presentation/page/register_page.dart';
-export 'feature/login/presentation/page/reset_password_page.dart';
+export 'feature/auth/presentation/page/login_page.dart';
+export 'feature/auth/presentation/page/register_page.dart';
+export 'feature/auth/presentation/page/reset_password_page.dart';
 export 'feature/memo/presentation/page/memo_page.dart';
 export 'feature/setting/presentation/page/setting_page.dart';
 export 'feature/user/presentation/page/account_page.dart';
